requirements.txt
streamlit
pyyaml
beautifulsoup4
pydantic